requirements.txt
asyncpg==0.22.0
databases[postgresql]==0.4.3
fastapi==0.63.0
psycopg2-binary==2.8.6
SQLAlchemy==1.3.24
uvicorn==0.13.4
gunicorn
# dev
pytest==6.2.3
requests==2.25.1
flake8
pep8
pytest-cov
python-dotenv==0.19.1